requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pydantic==1.10.8
openai==0.27.8
python-dotenv==0.21.1